+++ packages/components/avatar/src/avatar-group-context.ts
@@ -1,0 +1,8 @@
+import {createContext} from "@tw-material/react-utils";
+
+import {ContextType} from "./use-avatar-group";
+
+export const [AvatarGroupProvider, useAvatarGroupContext] = createContext<ContextType>({
+  name: "AvatarGroupContext",
+  strict: false,
+});