+++ packages/theme/src/components/button.ts
@@ -1,4 +1,5 @@
 import type {VariantProps} from "tailwind-variants";
+
 import {dataFocusVisibleClasses, tv} from "../utils";
 
 const button = tv({