+++ packages/theme/src/components/index.ts
@@ -4,3 +4,4 @@ export * from "./dropdown";
 export * from "./menu";
 export * from "./modal";
 export * from "./popover";
+export * from "./avatar";