+++ packages/components/dropdown/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @tw-material/dropdown
 
+## 1.1.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tw-material/popover@1.1.2
+  - @tw-material/menu@1.1.2
+
 ## 1.1.1
 
 ### Patch Changes