+++ packages/components/popover/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @tw-material/popover
 
+## 1.1.2
+
+### Patch Changes
+
+- changes
+
+- Updated dependencies []:
+  - @tw-material/button@1.1.2
+
 ## 1.1.1
 
 ### Patch Changes