+++ packages/theme/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @tw-material/theme
 
+## 1.1.2
+
+### Patch Changes
+
+- changes
+
 ## 1.1.1
 
 ### Patch Changes